Added command line options for program

+++ src/solver.cpp
@@ -5,8 +5,7 @@
 #include <iostream>
 
 
-Solver::Solver(std::shared_ptr<Board> board_ptr, std::string word_list_path, unsigned int min_word_length, unsigned int max_word_length)
-    : m_board_ptr(board_ptr), m_word_list_path(word_list_path), m_min_word_length(min_word_length), m_max_word_length(max_word_length)
+Solver::Solver(std::shared_ptr<Board> board_ptr, std::string word_list_path, int min_word_length, int max_word_length) : m_board_ptr(board_ptr), m_word_list_path(word_list_path), m_min_word_length(min_word_length), m_max_word_length(max_word_length)
 {
     Load_Word_List();
 
@@ -125,7 +124,7 @@ void Solver::Solve_From_Cell(unsigned int cell_x, unsigned int cell_y)
 
     while (direction_stack[0] != 8)
     {
-        if (current_word.size() == m_max_word_length || direction_stack.back() == 8)
+        if ((int)current_word.size() == m_max_word_length || direction_stack.back() == 8)
         {
             cells_visited[current_x][current_y] = false;
 
@@ -170,7 +169,7 @@ void Solver::Solve_From_Cell(unsigned int cell_x, unsigned int cell_y)
             direction_stack.push_back(0);
             search_range_stack.push_back(range);
 
-            if (current_word.size() >= m_min_word_length && m_found_words.find(current_word) == m_found_words.end())
+            if ((int)current_word.size() >= m_min_word_length && m_found_words.find(current_word) == m_found_words.end())
             {
                 for (int i = range.first; i <= range.second; i++)
                 {
